- Map Screen: Displaying Alarm (+ circle Distances of Notify), Create Alarm on Long tap on map, Move Alarm marker on map (change position) - Fix Location Service starting when allow permissions - Fix Create/Edit Location Alarm logic

+++ app/src/main/java/com/vladbakalo/location_alarm/data/models/Alarm.kt
@@ -10,7 +10,7 @@ data class Alarm(
     @PrimaryKey(autoGenerate = true) val id: Long,
     var enabled: Boolean,
     @ColumnInfo(name = "location_alarm_id") var locationAlarmId: Long,
-    @ColumnInfo(name = "notify_distance") var notifyDistanceMeters: Double) : BaseDataModel {
+    @ColumnInfo(name = "notify_distance") var notifyDistanceMeters: Int) : BaseDataModel {
     override fun getPrimaryId(): Long {
         return id
     }
